migrate TestProposalQueues to use simapp
